+++ app/Http/Controllers/PatientController.php
@@ -21,38 +21,6 @@ class PatientController extends Controller
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
@@ -97,8 +65,13 @@ class PatientController extends Controller
     {
         $patient = Patient::findOrFail($id);
 
-        $patient->delete();
+        if(!count($patient->appointments)) {
 
-        return redirect('/patients')->with('msg', 'Patient deleted successfully!');
+            $patient->delete();
+
+            return redirect('/patients')->with('msg', 'Patient deleted successfully!');
+        }
+
+        return redirect('patients')->with('msg', "This patient has appointments. Can't delete it");
     }
 }
